# Optimize delta decoding (part 2)