--- a/encoding/delta/binary_packed_amd64.go
+++ b/encoding/delta/binary_packed_amd64.go
@@ -14,6 +14,7 @@ func init() {
 
 		decodeMiniBlockInt32Table[0] = decodeMiniBlockInt32x1bitAVX2
 		decodeMiniBlockInt32Table[1] = decodeMiniBlockInt32x2bitsAVX2
+		decodeMiniBlockInt32Table[3] = decodeMiniBlockInt32x4bitsAVX2
 		decodeMiniBlockInt32Table[7] = decodeMiniBlockInt32x8bitsAVX2
 		decodeMiniBlockInt32Table[15] = decodeMiniBlockInt32x16bitsAVX2
 	}
@@ -215,6 +216,9 @@ func decodeMiniBlockInt32x1bitAVX2(dst []int32, src []uint32)
 
 //go:noescape
 func decodeMiniBlockInt32x2bitsAVX2(dst []int32, src []uint32)
+
+//go:noescape
+func decodeMiniBlockInt32x4bitsAVX2(dst []int32, src []uint32)
 
 //go:noescape
 func decodeMiniBlockInt32x8bitsAVX2(dst []int32, src []uint32)