--- a/encoding/delta/binary_packed_amd64.go
+++ b/encoding/delta/binary_packed_amd64.go
@@ -27,36 +27,36 @@ func blockSubInt32AVX2(block *[blockSize]int32, value int32)
 func blockBitWidthsInt32AVX2(bitWidths *[numMiniBlocks]byte, block *[blockSize]int32)
 
 //go:noescape
-func miniBlockPackInt32Default(dst *byte, src *[miniBlockSize]int32, bitWidth uint)
+func encodeMiniBlockInt32Default(dst *byte, src *[miniBlockSize]int32, bitWidth uint)
 
 //go:noescape
-func miniBlockPackInt32x1bitAVX2(dst *byte, src *[miniBlockSize]int32)
+func encodeMiniBlockInt32x1bitAVX2(dst *byte, src *[miniBlockSize]int32)
 
 //go:noescape
-func miniBlockPackInt32x2bitsAVX2(dst *byte, src *[miniBlockSize]int32)
+func encodeMiniBlockInt32x2bitsAVX2(dst *byte, src *[miniBlockSize]int32)
 
 //go:noescape
-func miniBlockPackInt32x3to16bitsAVX2(dst *byte, src *[miniBlockSize]int32, bitWidth uint)
+func encodeMiniBlockInt32x3to16bitsAVX2(dst *byte, src *[miniBlockSize]int32, bitWidth uint)
 
 //go:noescape
-func miniBlockPackInt32x32bitsAVX2(dst *byte, src *[miniBlockSize]int32)
+func encodeMiniBlockInt32x32bitsAVX2(dst *byte, src *[miniBlockSize]int32)
 
-func miniBlockPackInt32(dst []byte, src *[miniBlockSize]int32, bitWidth uint) {
-	miniBlockPackInt32Default(&dst[0], src, bitWidth)
+func encodeMiniBlockInt32(dst []byte, src *[miniBlockSize]int32, bitWidth uint) {
+	encodeMiniBlockInt32Default(&dst[0], src, bitWidth)
 }
 
-func miniBlockPackInt32AVX2(dst *byte, src *[miniBlockSize]int32, bitWidth uint) {
+func encodeMiniBlockInt32AVX2(dst *byte, src *[miniBlockSize]int32, bitWidth uint) {
 	switch {
 	case bitWidth == 1:
-		miniBlockPackInt32x1bitAVX2(dst, src)
+		encodeMiniBlockInt32x1bitAVX2(dst, src)
 	case bitWidth == 2:
-		miniBlockPackInt32x2bitsAVX2(dst, src)
+		encodeMiniBlockInt32x2bitsAVX2(dst, src)
 	case bitWidth == 32:
-		miniBlockPackInt32x32bitsAVX2(dst, src)
+		encodeMiniBlockInt32x32bitsAVX2(dst, src)
 	case bitWidth <= 16:
-		miniBlockPackInt32x3to16bitsAVX2(dst, src, bitWidth)
+		encodeMiniBlockInt32x3to16bitsAVX2(dst, src, bitWidth)
 	default:
-		miniBlockPackInt32Default(dst, src, bitWidth)
+		encodeMiniBlockInt32Default(dst, src, bitWidth)
 	}
 }
 
@@ -95,7 +95,7 @@ func encodeInt32AVX2(dst []byte, src []int32) []byte {
 		for i, bitWidth := range bitWidths {
 			if bitWidth != 0 {
 				miniBlock := (*[miniBlockSize]int32)(block[i*miniBlockSize:])
-				miniBlockPackInt32AVX2(&dst[n], miniBlock, uint(bitWidth))
+				encodeMiniBlockInt32AVX2(&dst[n], miniBlock, uint(bitWidth))
 				n += (miniBlockSize * int(bitWidth)) / 8
 			}
 		}
@@ -119,31 +119,31 @@ func blockSubInt64AVX2(block *[blockSize]int64, value int64)
 func blockBitWidthsInt64AVX2(bitWidths *[numMiniBlocks]byte, block *[blockSize]int64)
 
 //go:noescape
-func miniBlockPackInt64Default(dst *byte, src *[miniBlockSize]int64, bitWidth uint)
+func encodeMiniBlockInt64Default(dst *byte, src *[miniBlockSize]int64, bitWidth uint)
 
 //go:noescape
-func miniBlockPackInt64x1bitAVX2(dst *byte, src *[miniBlockSize]int64)
+func encodeMiniBlockInt64x1bitAVX2(dst *byte, src *[miniBlockSize]int64)
 
 //go:noescape
-func miniBlockPackInt64x2bitsAVX2(dst *byte, src *[miniBlockSize]int64)
+func encodeMiniBlockInt64x2bitsAVX2(dst *byte, src *[miniBlockSize]int64)
 
 //go:noescape
-func miniBlockPackInt64x64bitsAVX2(dst *byte, src *[miniBlockSize]int64)
+func encodeMiniBlockInt64x64bitsAVX2(dst *byte, src *[miniBlockSize]int64)
 
-func miniBlockPackInt64(dst []byte, src *[miniBlockSize]int64, bitWidth uint) {
-	miniBlockPackInt64Default(&dst[0], src, bitWidth)
+func encodeMiniBlockInt64(dst []byte, src *[miniBlockSize]int64, bitWidth uint) {
+	encodeMiniBlockInt64Default(&dst[0], src, bitWidth)
 }
 
-func miniBlockPackInt64AVX2(dst *byte, src *[miniBlockSize]int64, bitWidth uint) {
+func encodeMiniBlockInt64AVX2(dst *byte, src *[miniBlockSize]int64, bitWidth uint) {
 	switch {
 	case bitWidth == 1:
-		miniBlockPackInt64x1bitAVX2(dst, src)
+		encodeMiniBlockInt64x1bitAVX2(dst, src)
 	case bitWidth == 2:
-		miniBlockPackInt64x2bitsAVX2(dst, src)
+		encodeMiniBlockInt64x2bitsAVX2(dst, src)
 	case bitWidth == 64:
-		miniBlockPackInt64x64bitsAVX2(dst, src)
+		encodeMiniBlockInt64x64bitsAVX2(dst, src)
 	default:
-		miniBlockPackInt64Default(dst, src, bitWidth)
+		encodeMiniBlockInt64Default(dst, src, bitWidth)
 	}
 }
 
@@ -182,7 +182,7 @@ func encodeInt64AVX2(dst []byte, src []int64) []byte {
 		for i, bitWidth := range bitWidths {
 			if bitWidth != 0 {
 				miniBlock := (*[miniBlockSize]int64)(block[i*miniBlockSize:])
-				miniBlockPackInt64AVX2(&dst[n], miniBlock, uint(bitWidth))
+				encodeMiniBlockInt64AVX2(&dst[n], miniBlock, uint(bitWidth))
 				n += (miniBlockSize * int(bitWidth)) / 8
 			}
 		}