Following up from https://github.com/segmentio/parquet-go/pull/229, this PR contributes to https://github.com/segmentio/parquet-go/issues/226 by modifying the decoding of mini blocks in delta encoding using the techniques described in https://arxiv.org/pdf/1209.2137v5.pdf

I found these parts of the paper to be the best explanation of how the bit unpacking algorithm works:

<img width="1020" alt="image" src="https://user-images.githubusercontent.com/865510/173222320-9546112f-e9fe-43c5-8c63-e3cb1a049916.png">

> Figure 8. One step of simultaneous decoding of four 5-bit integers that are stored in a horizontal layout (as
opposed to the vertical data layout of Fig. 6). These integers are copied to four 32-bit words using the shuffle
operation pshufb. The locations in source and target buffers are indicated by arrows. Curvy lines are used
to denote integers that cross byte boundaries in the source buffer. Hence, they are copied only partially. The
boldface zero values represent the bytes zeroed by the shuffle instruction. Note that some source bytes are
copied to multiple locations.

_Page 13_

The implementation was quite complex, I iterated over it multiple times, starting with simple specializations and discovering how to make the algorithm more generic over time. I left comments in the code to capture some of the learning I thought could be relevant to future readers.

The following snapshots of benchmarks illustrate the impact of the change:
```
name                              old time/op    new time/op    delta
Decode/DELTA_BINARY_PACKED/int32    19.7µs ± 0%     9.4µs ± 0%   -52.16%  (p=0.000 n=10+10)

name                              old speed      new speed      delta
Decode/DELTA_BINARY_PACKED/int32  2.03GB/s ± 0%  4.25GB/s ± 0%  +109.04%  (p=0.000 n=10+10)

name                              old value/s    new value/s    delta
Decode/DELTA_BINARY_PACKED/int32      509M ± 0%     1063M ± 0%  +109.04%  (p=0.000 n=10+10)
```
```
name                                  old time/op  new time/op  delta
MergeFiles/INT32/groups=3,rows=30000  8.27µs ± 1%  7.11µs ± 2%  -14.05%  (p=0.000 n=10+9)
MergeFiles/UUID/groups=3,rows=30000   20.2µs ± 1%  18.4µs ± 1%   -9.32%  (p=0.000 n=10+9)

name                                  old row/s    new row/s    delta
MergeFiles/INT32/groups=3,rows=30000    119M ± 1%    138M ± 2%  +16.35%  (p=0.000 n=10+9)
MergeFiles/UUID/groups=3,rows=30000    48.6M ± 1%   53.6M ± 1%  +10.28%  (p=0.000 n=10+9)
```

Please take a look and let me know if you think anything could be improved.